requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.3
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.8.2
billiard==4.2.1
bs4==0.0.2
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
compressed-rtf==1.0.7
contourpy==1.3.2
cycler==0.12.1
docx2txt==0.9
ebcdic==1.1.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
extract-msg==0.28.7
fastapi==0.115.12
fonttools==4.57.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
IMAPClient==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.8
kombu==5.5.3
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==2.2.5
olefile==0.47
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdfminer.six==20191110
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==0.6.23
pytz==2025.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.8.1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.14.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
xlrd==1.2.0
XlsxWriter==3.2.3
yarl==1.20.0
pdfplumber==0.10.3
